Support Vulkan on Android
Since atmoic support isn't guaranteed on Vulkan, add a new PlatformFeature called 'supportsFragmentShaderAtomics'.

Various bugfixes.

Download and package Vulkan validation layers with the android testing apk.

Start an android Vulkan runner that doesn't do diffing yet.

Diffs=
cc903ee02 Support Vulkan on Android (#7915)

+++ renderer/src/shaders/glsl.glsl
@@ -338,7 +338,7 @@
     layout(location = IDX) out lowp vec4 NAME
 #define PLS_DECLUI(IDX, NAME)                                                                      \
     layout(input_attachment_index = IDX, binding = IDX, set = PLS_TEXTURE_BINDINGS_SET)            \
-        uniform lowp usubpassInput _in_##NAME;                                                     \
+        uniform highp usubpassInput _in_##NAME;                                                    \
     layout(location = IDX) out highp uvec4 NAME
 #define PLS_BLOCK_END
